refactor: move RequestError from network to interfaces p2p
