+++ crates/interfaces/src/p2p/mod.rs
@@ -6,6 +6,10 @@
 /// [`HeadersClient`]: crate::p2p::headers::HeadersClient
 pub mod headers;
 
+/// Error types broadly used by p2p interfaces for any operation which may produce an error when
+/// interacting with the network implementation
+pub mod error;
+
 use futures::Stream;
 use std::pin::Pin;
