+++ crates/net/network/src/message.rs
@@ -13,37 +13,9 @@ use std::task::{ready, Context, Poll};
 
 use crate::NodeId;
 use reth_eth_wire::capability::CapabilityMessage;
+use reth_interfaces::p2p::error::RequestResult;
 use reth_primitives::{Header, Receipt, TransactionSigned};
 use tokio::sync::{mpsc, mpsc::error::TrySendError, oneshot};
-
-/// Result alias for result of a request.
-pub type RequestResult<T> = Result<T, RequestError>;
-
-/// Error variants that can happen when sending requests to a session.
-#[derive(Debug, thiserror::Error)]
-#[allow(missing_docs)]
-pub enum RequestError {
-    #[error("Closed channel.")]
-    ChannelClosed,
-    #[error("Not connected to the node.")]
-    NotConnected,
-    #[error("Capability Message is not supported by remote peer.")]
-    UnsupportedCapability,
-    #[error("Network error: {0}")]
-    Io(String),
-}
-
-impl<T> From<mpsc::error::SendError<T>> for RequestError {
-    fn from(_: mpsc::error::SendError<T>) -> Self {
-        RequestError::ChannelClosed
-    }
-}
-
-impl From<oneshot::error::RecvError> for RequestError {
-    fn from(_: oneshot::error::RecvError) -> Self {
-        RequestError::ChannelClosed
-    }
-}
 
 /// Represents all messages that can be sent to a peer session
 #[derive(Debug)]
